# output/openapi.yaml
openapi: 3.0.0
info:
  description: 会员系统前后端分离接口文档
  title: 会员系统 API
  version: 1.0.0
servers:
- url: http://localhost:8000
...
paths:
...
  /api/notifications:
    get:
      responses:
        "200":
          description: 通知列表
      security:
      - bearerAuth: []
      summary: 获取系统通知列表
      tags:
      - 通知
  /api/notifications/{id}/read:
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          description: 标记成功
      security:
      - bearerAuth: []
      summary: 标记通知为已读
      tags:
      - 通知
...
components:
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http